requirements.txt
amqp==5.2.0
arrow==1.3.0
asgiref==3.8.1
async-timeout==4.0.3
billiard==4.2.0
celery==5.4.0
certifi==2024.6.2
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.3
Django==4.2.13
django-admin-inline-paginator-plus==0.1.1
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-cors-headers==4.3.1
django-filter==24.2
django-js-asset==2.2.0
django-jsonform==2.22.0
django-model-utils==4.5.1
django-mptt==0.16.0
django-redis==5.4.0
django-timezone-field==7.0
djangorestframework==3.15.1
djangorestframework-recursive==0.1.2
djangorestframework-simplejwt==5.3.1
drf-extensions==0.7.1
drf-yasg==1.21.7
filelock==3.16.0
gunicorn==23.0.0
idna==3.7
inflection==0.5.1
Jinja2==3.1.4
kombu==5.3.7
Markdown==3.6
MarkupSafe==2.1.5
optionaldict==0.1.2
packaging==24.1
ply==3.11
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
pycparser==2.22
pycryptodome==3.20.0
PyJWT==2.8.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
redis==5.0.7
requests==2.32.3
requests-file==2.1.0
rule-engine==4.5.0
sentry-sdk==2.9.0
six==1.16.0
sqlparse==0.5.0
tenacity==9.0.0
tldextract==5.1.2
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.1
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
uWSGI==2.0.26
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.7.0
wrapt==1.16.0
xmltodict==0.13.0
